+++ src/xrGame/attachment_owner.cpp
@@ -187,7 +187,7 @@ CAttachableItem* CAttachmentOwner::attachedItem(CLASS_ID clsid) const
 	return NULL;
 }
 
-CAttachableItem* CAttachmentOwner::attachedItem(u16 id) const
+CAttachableItem* CAttachmentOwner::attachedItem(u32 id) const
 {
 	xr_vector<CAttachableItem*>::const_iterator I = m_attached_objects.begin();
 	xr_vector<CAttachableItem*>::const_iterator E = m_attached_objects.end();
