+++ src/xrGame/attachment_owner.h
@@ -37,7 +37,7 @@ public:
 	IC const xr_vector<CAttachableItem*>& attached_objects() const;
 
 	CAttachableItem* attachedItem(CLASS_ID clsid) const;
-	CAttachableItem* attachedItem(u16 id) const;
+	CAttachableItem* attachedItem(u32 id) const;
 	CAttachableItem* attachedItem(shared_str& section) const;
 };
